created index route
