+++ package.json
@@ -11,6 +11,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "ejs": "^2.5.8",
     "express": "^4.15.3"
   }
 }
